move evaluate_convergence to htsohm
